+++ Src/core/mat_nos_math.cpp
@@ -456,7 +456,7 @@ void get_submatrix(const T2& full,
     err_checkf(indices.size() > 0, "Atom indices list is empty.", std::cout);
     const int n = static_cast<int>(indices.size());
     err_checkf(full.extent(0) == full.extent(1), "Matrix must be square.", std::cout);
-    err_checkf(sub.size() == n * n, "Submatrix has incorrect size.", std::cout);
+    err_checkf(sub.size() == static_cast<size_t>(n) * n, "Submatrix has incorrect size.", std::cout);
 
     for (int i = 0; i < n; ++i) {
         const int global_i = indices[i];
@@ -480,7 +480,7 @@ void get_submatrix(const T2& full,
     const int n2 = static_cast<int>(vec_indices.size());
     err_checkf(n1 > 0, "Val indices list is empty.", std::cout);
     err_checkf(n2 > 0, "Vec indices list is empty.", std::cout);
-    err_checkf(sub.size() == n1 * n2, "Submatrix has incorrect size.", std::cout);
+    err_checkf(sub.size() == static_cast<size_t>(n1) * n2, "Submatrix has incorrect size.", std::cout);
 
     for (int i = 0; i < n1; ++i) {
         const int global_i = val_indices[i];
@@ -583,8 +583,8 @@ void get_submatrices(const T2& D_full,
     err_checkf(D_full.extent(0) == D_full.extent(1), "Density matrix D must be square.", std::cout);
     err_checkf(S_full.extent(0) == S_full.extent(1), "Overlap matrix S must be square.", std::cout);
     err_checkf(D_full.extent(0) == S_full.extent(0), "Density and Overlap matrices must be of the same size.", std::cout);
-    err_checkf(D_sub.size() == n * n, "Density submatrix has incorrect size.", std::cout);
-    err_checkf(S_sub.size() == n * n, "Overlap submatrix has incorrect size.", std::cout);
+    err_checkf(D_sub.size() == static_cast<size_t>(n) * n, "Density submatrix has incorrect size.", std::cout);
+    err_checkf(S_sub.size() == static_cast<size_t>(n) * n, "Overlap submatrix has incorrect size.", std::cout);
 
     for (int i = 0; i < n; ++i) {
         const int global_i = indices[i];
@@ -607,8 +607,8 @@ dMatrix2 LAPACKE_invert(const dMatrix2& A, const double cutoff) {
 
     // 1. Allocate memory for SVD results
     vec S(k);                 // Singular values
-    vec U(m * k);             // Left singular vectors (m x k)
-    vec Vt(k * n);            // Right singular vectors transposed (k x n)
+    vec U(static_cast<size_t>(m) * k);             // Left singular vectors (m x k)
+    vec Vt(static_cast<size_t>(k) * n);            // Right singular vectors transposed (k x n)
     vec superb(k - 1);        // Workspace for dgesvd
 
     // Make a copy of A because dgesvd destroys the input matrix
@@ -680,7 +680,7 @@ dMatrix2 LAPACKE_invert(const dMatrix2& A, const double cutoff) {
         S[i] = S[i] < cutoff ? 0.0 : 1.0 / S[i];
 
     // 4. Compute A^+ = V * S^+ * U^T
-    A_copy = vec(k * m, 0.0); // Reuse A_copy as W to save memory
+    A_copy = vec(static_cast<size_t>(k) * m, 0.0); // Reuse A_copy as W to save memory
     for (int i = 0; i < k; ++i) {
         const int im = i * m;
         if (S[i] == 0.0)
@@ -741,7 +741,7 @@ void make_Eigenvalues(vec& A, vec& W) {
 
 vec mat_sqrt(vec& A, vec& W, const double cutoff) {
     const int n = static_cast<int>(W.size());
-    vec Temp(n * n, 0.0);
+    vec Temp(static_cast<size_t>(n) * n, 0.0);
 
     make_Eigenvalues(A, W);
